requirements.txt
Flask==2.0.1
Flask-PyMongo==2.3.0
python-dotenv==0.19.0
pandas==1.1.3
openpyxl==3.0.9
flask-cors==3.0.10